[2/2] Moves remaining files from controllers to resolvers file
